# server/docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreateExerciseRequest:
    properties:
      category:
        enum:
        - strength
        - cardio
        - bodyweight
        - stretch
        type: string
      equipment:
        enum:
        - barbell
        - dumbbell
        - machine
        - cable
        - bodyweight
        - band
        - kettlebell
        - ez_bar
        - smith_machine
        - other
        type: string
      name:
        maxLength: 100
        minLength: 1
        type: string
      name_en:
        maxLength: 100
        type: string
      notes:
        maxLength: 500
        type: string
      primary_muscles:
        items:
          type: string
        minItems: 1
        type: array
      secondary_muscles:
        items:
          type: string
        type: array
      tracking_type:
        enum:
        - weight_reps
        - time
        - distance
        - reps_only
        type: string
    required:
    - category
    - equipment
    - name
    - primary_muscles
    - tracking_type
    type: object
...
  dto.ExerciseResponse:
    properties:
      category:
        type: string
      created_at:
        type: string
      equipment:
        type: string
      id:
        type: string
      is_custom:
        type: boolean
      name:
        type: string
      name_en:
        type: string
      notes:
        type: string
      primary_muscles:
        items:
          type: string
        type: array
      secondary_muscles:
        items:
          type: string
        type: array
      tracking_type:
        type: string
    type: object
...
  dto.UpdateExerciseRequest:
    properties:
      category:
        enum:
        - strength
        - cardio
        - bodyweight
        - stretch
        type: string
      equipment:
        type: string
      name:
        maxLength: 100
        minLength: 1
        type: string
      name_en:
        maxLength: 100
        type: string
      notes:
        maxLength: 500
        type: string
      primary_muscles:
        items:
          type: string
        minItems: 1
        type: array
      secondary_muscles:
        items:
          type: string
        type: array
      tracking_type:
        type: string
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
info:
  contact: {}
  description: Backend service for the strength training record app.
  title: Strength Tracker API
  version: "1.0"
paths:
...
  /exercises:
    get:
      parameters:
      - description: Exercise category
        in: query
        name: category
        type: string
      - description: Primary muscle
        in: query
        name: muscle
        type: string
      - description: Equipment
        in: query
        name: equipment
        type: string
      - description: Search by name or name_en
        in: query
        name: search
        type: string
      - description: Only custom or preset exercises
        in: query
        name: is_custom
        type: boolean
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      - description: Sort field
        in: query
        name: sort_by
        type: string
      - description: Sort direction
        in: query
        name: sort_dir
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: List exercises
      tags:
      - exercises
    post:
      consumes:
      - application/json
      parameters:
      - description: Create exercise request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateExerciseRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.ExerciseResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Create custom exercise
      tags:
      - exercises
  /exercises/{id}:
    delete:
      parameters:
      - description: Exercise ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Archive custom exercise
      tags:
      - exercises
    get:
      parameters:
      - description: Exercise ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.ExerciseResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get exercise detail
      tags:
      - exercises
    put:
      consumes:
      - application/json
      parameters:
      - description: Exercise ID
        in: path
        name: id
        required: true
        type: string
      - description: Update exercise request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateExerciseRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.ExerciseResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Update custom exercise
      tags:
      - exercises
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"